Add build_img.sh helper to build sdcard images
We plan on A/B booting our fika based boards which is not supported
by variscites debian.
The added script therefore creates a new partition scheme either for
an image or for an already existing SDCard and unpacks the rootfs
into it.
For the moment we are using the precompressed tarbal as it preserves
the permissions but testing if rsync would also work could massively
improve development turnaround times in the future.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>