var: Timeout wifi service after 20 seconds
The variscite wifi service is to be described flimsy at best.
Not only does it rely on the deprecated CONFIG_SYSFS_GPIO but also
does load the kernel module brcmfm which must not be compiled into
the kernel but has be to existing as a module.
During development many of these requirements can break leading to
a failing service.
Don't block the systems boot by timeouting the service after a generous
amount of time.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>