u-boot: Update bootloader to variscites 2022.04
Testing revealed that we can use the 2022.04 u-boot as a
drop-in replacement for the 2020.04 that variscite ships
with debian and its original 5.4 kernel.
The 5.15 kernel we are now running as well as the newly released
6.1 kernel both boot fine with that 2022.04 u-boot.

Sadly we still have to rely on the NXP mkimage tool as the binman
image generation is not yet working in this u-boot.

Therefore for flashing we run:
sudo ./var_make_debian.sh -c bootloader && sudo dd if=output/imx-boot-sd.bin of=/dev/sdc bs=1K seek=32

Signed-off-by: Mimoja <mimoja@meticuloushome.com>